Moving istainted as well
