+++ manticore/ethereum/detectors.py
@@ -3,8 +3,7 @@ import hashlib
 import logging
 from contextlib import contextmanager
 
-from ..core.smtlib import Operators, Constant, simplify, issymbolic
-from ..utils.helpers import istainted, taint_with, get_taints
+from ..core.smtlib import Operators, Constant, simplify, issymbolic, istainted, get_taints, taint_with
 from ..core.plugin import Plugin
