+++ manticore/platforms/evm.py
@@ -6,7 +6,6 @@ import copy
 import inspect
 from functools import wraps
 from typing import List, Set, Tuple, Union
-from ..utils.helpers import get_taints, taint_with, istainted
 from ..platforms.platform import *
 from ..core.smtlib import (
     Z3Solver,
@@ -22,6 +21,9 @@ from ..core.smtlib import (
     to_constant,
     simplify,
     issymbolic,
+    get_taints,
+    istainted,
+    taint_with,
 )
 from ..core.state import Concretize, TerminateState
 from ..utils.event import Eventful
